requirements.txt
bibtexparser>=1.4.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.0.0
tqdm>=4.65.0
beautifulsoup4>=4.12.0
requests>=2.31.0